requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
jinja2
python-multipart
langchain
faiss-cpu
openai
langchain-community
langchain-groq
sentence-transformers
tiktoken
itsdangerous
python-dotenv
langchain_huggingface
python-docx 
pypdf 
pandas 
openpyxl 
unstructured
doc2txt
markdown
markdown
bleach
pygments

